--- openmind/src/Components/InputTextArea.jsx
@@ -1,7 +1,0 @@
-import Styles from '../Styles/InputTextArea.module.css';
-
-const InputTextArea = () => {
-  return <input className={Styles.input} type='text' placeholder='이름을 입력하세요' />;
-};
-
-export default InputTextArea;
